store Info form 수정

+++ teamProject/src/main/java/com/spring/teamProject/vo/ImageFileVO.java
@@ -44,7 +44,7 @@ public class ImageFileVO {
 		return reviewId;
 	}
 
-	public void setReviewId(int reviewId) {
+	public void setReviewId(long reviewId) {
 		this.reviewId = reviewId;
 	}
